ajustes nas migrates e criação de seedes

+++ database/migrations/2024_01_09_173749_create_jogo_plataforma_table.php
@@ -13,15 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('jogo_plataforma', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('jogo_id');
-            $table->unsignedBigInteger('plataforma_id');
-            $table->timestamps();
+        // Verifica se a tabela 'jogo_plataforma' já existe antes de criar
+        if (!Schema::hasTable('jogo_plataforma')) {
+            Schema::create('jogo_plataforma', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('jogo_id');
+                $table->unsignedBigInteger('plataforma_id');
+                $table->timestamps();
 
-            $table->foreign('jogo_id')->references('id')->on('jogos')->onDelete('cascade');
-            $table->foreign('plataforma_id')->references('id')->on('plataformas')->onDelete('cascade');
-        });
+                $table->foreign('jogo_id')->references('id')->on('jogos')->onDelete('cascade');
+                $table->foreign('plataforma_id')->references('id')->on('plataformas')->onDelete('cascade');
+            });
+        }
     }
 
     /**
